backend/backend_requirements.txt
django==4.2.7
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
selenium==4.15.2
webdriver-manager==4.0.1
scikit-learn==1.3.2
nltk==3.8.1
django-redis==5.4.0
psycopg2-binary==2.9.9
djangorestframework==3.14.0